requirements.txt
python-telegram-bot[job-queue]==22.6
python-dotenv==1.2.1
aiohttp
requests==2.32.5
urllib3==2.6.2
yarl==1.22.0
sniffio==1.3.1
beautifulsoup4
soupsieve==2.8
lxml
Pillow
pytesseract==0.3.13
psutil
python-dateutil==2.9.0.post0
six==1.17.0
python-whois==0.9.6
yt-dlp-ejs
deep-translator
dnspython
uvloop==0.22.1
aiofiles
yt-dlp
